ai-agents/requirements.txt
letta-client>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.31.0
aiohttp>=3.8.0
flask>=2.0.0
flask-cors>=4.0.0
